refactor run_code_ut and create run_code.py to enabale project run normally
